split up schema fields into submodules
